[COMMON] mmc: dw_mmc: fix type variable for prevent detect.

+++ drivers/mmc/host/dw_mmc.h
@@ -383,8 +383,8 @@ struct dw_mci_req_log {
 	u32				info1;
 	u32				info2;
 	u32				info3;
-	u32				pending_events;
-	u32				completed_events;
+	unsigned long			pending_events;
+	unsigned long			completed_events;
 	enum dw_mci_state		state;
 	enum dw_mci_state		state_cmd;
 	enum dw_mci_state		state_dat;
@@ -502,8 +502,8 @@ struct dw_mci_sfe_ram_dump {
 	u32			mpsctrl;
 	u32			cmd_status;
 	u32			data_status;
-	u32			pending_events;
-	u32			completed_events;
+	unsigned long		pending_events;
+	unsigned long		completed_events;
 	u32			host_state;
 	u32			cmd_index;
 	u32			fifo_count;
